Detailed Documentation - ML
Writing detailed notes and code on a jupyter notebook simultaneously doesn't seem to embed the knowledge _that_ well for me.

Brief notes, sure, but I want detailed descriptions to really embed what I'm learning. The purpose of this documentation is to do just that - to write up detailed notes on all of my ML-related learning.